requirements.txt
redis
websocket-client
requests